simplified negotiation phases:  - state machine  - hello buffers were moved to mw1  - ntrdma is managing the hello buffers instead of ntc_ntb_msi (i.e double buffering etc)

+++ drivers/ntc/ntc.c
@@ -142,7 +142,7 @@ void ntc_clear_ctx(struct ntc_dev *ntc)
 }
 EXPORT_SYMBOL(ntc_clear_ctx);
 
-ssize_t ntc_ctx_hello(struct ntc_dev *ntc, int phase,
+int ntc_ctx_hello(struct ntc_dev *ntc, int phase,
 		      void *in_buf, size_t in_size,
 		      void *out_buf, size_t out_size)
 {
